First pass on Evan's PR
This is the first pass for integrating Evan's PR in the main code.
Some functionalities have been disabled and some more have beed altered, but this reproduces the 6j and the 3H selections implemented in the original PR.
Running on the same 6b signal events, a perfect match between the original PR and this code was found.